+++ lib/temple/parsers/void_elements_aliases.ex
@@ -1,0 +1,29 @@
+defmodule Temple.Parser.VoidElementsAliases do
+  @behaviour Temple.Parser
+
+  alias Temple.Parser
+  alias Temple.Buffer
+
+  @impl Parser
+  def applicable?({name, _, _}) do
+    name in Parser.void_elements_aliases()
+  end
+
+  def applicable?(_), do: false
+
+  @impl Parser
+  def run({name, _, args}, buffer) do
+    import Temple.Parser.Private
+
+    {_do_and_else, args} =
+      args
+      |> split_args()
+
+    name = Parser.void_elements_lookup()[name]
+
+    Buffer.put(buffer, "<#{name}#{compile_attrs(args)}>")
+    Buffer.put(buffer, "\n")
+
+    :ok
+  end
+end
